Version 1
This version isn't really working as I would like, it is more of a proof of concept than a working game.
I think I will need to make a new branch where I convert all the 3D RigidBody componentsw to 2D RigidBody components, it seems to me as if we the game can be a 2D game.
My main issue is to do with the ISO camera not being able to target or shoot at objects in 3D space.